# Replace object spread a reduce in generateCSSRuleset