--- a/src/generate.js
+++ b/src/generate.js
@@ -250,6 +250,11 @@ const transformRule = (
 );
 
 
+const arrayToObjectKeysReducer = (acc, val) => {
+    acc[val] = true;
+    return acc;
+};
+
 /**
  * Generate a CSS ruleset with the selector and containing the declarations.
  *
@@ -291,7 +296,8 @@ export const generateCSSRuleset = (
     // Mutates declarations
     runStringHandlers(declarations, stringHandlers, selectorHandlers);
 
-    const originalElements = new Set(Object.keys(declarations.elements));
+    const originalElements = Object.keys(declarations.elements)
+        .reduce(arrayToObjectKeysReducer, Object.create(null));
 
     // NOTE(emily): This mutates handledDeclarations.elements.
     const prefixedElements = prefixAll(declarations.elements);
@@ -305,7 +311,7 @@ export const generateCSSRuleset = (
         // sortOrder, which means it was added by prefixAll. This means that we
         // need to figure out where it should appear in the sortOrder.
         for (let i = 0; i < elementNames.length; i++) {
-            if (!originalElements.has(elementNames[i])) {
+            if (!originalElements[elementNames[i]]) {
                 // This element is not in the sortOrder, which means it is a prefixed
                 // value that was added by prefixAll. Let's try to figure out where it
                 // goes.
@@ -324,7 +330,7 @@ export const generateCSSRuleset = (
                     originalStyle = elementNames[i][2].toLowerCase() + elementNames[i].slice(3);
                 }
 
-                if (originalStyle && originalElements.has(originalStyle)) {
+                if (originalStyle && originalElements[originalStyle]) {
                     const originalIndex = declarations.keyOrder.indexOf(originalStyle);
                     declarations.keyOrder.splice(originalIndex, 0, elementNames[i]);
                 } else {