I was profiling Aphrodite and noticed that the self time in
generateCSSRuleset is one of the most expensive bits. Digging in a bit,
it looked like most of this was coming from the object spread here,
which we use do copy the original object so we can reference its keys
with fast lookups. By swapping this out for a ~Set~ reduce we greatly reduce the
self time of this function from ~7% of total runtime to ~3%.

~I've implemented this in a way that expects Set to be available in the
browser or polyfilled, which I think is a reasonable explanation. I've
documented this in the readme. If we move forward with this, we could
also make Map a requirement and simplify a couple of codepaths if we
want.~