@lencioni 2018-02-16
@xymostech I think this would be worth including in the next release 💃