@lencioni 2018-02-16
I ran this through https://github.com/hellofresh/css-in-js-perf-tests and here are my results.

## master

```
Running simple test.

  aphrodite              x 30,105 ops/sec ±1.28% (81 runs sampled)
  aphrodite-no-important x 31,924 ops/sec ±0.94% (86 runs sampled)

Running nested test.

  aphrodite              x 12,276 ops/sec ±2.35% (85 runs sampled)
  aphrodite-no-important x 13,814 ops/sec ±1.48% (87 runs sampled)

Running style overload test.

  aphrodite              x 3,991 ops/sec ±4.13% (80 runs sampled)
  aphrodite-no-important x 4,059 ops/sec ±12.48% (77 runs sampled)

Running classes overload test.

  aphrodite              x 4,526 ops/sec ±3.02% (78 runs sampled)
  aphrodite-no-important x 5,095 ops/sec ±4.26% (79 runs sampled)

```

## This branch

```
Running simple test.

  aphrodite              x 29,719 ops/sec ±1.22% (80 runs sampled)
  aphrodite-no-important x 32,488 ops/sec ±1.16% (87 runs sampled)


Running nested test.

  aphrodite              x 12,535 ops/sec ±1.55% (82 runs sampled)
  aphrodite-no-important x 14,301 ops/sec ±0.93% (89 runs sampled)


Running style overload test.

  aphrodite              x 4,153 ops/sec ±1.80% (81 runs sampled)
  aphrodite-no-important x 4,631 ops/sec ±1.62% (86 runs sampled)

Running classes overload test.

  aphrodite              x 5,242 ops/sec ±1.28% (83 runs sampled)
  aphrodite-no-important x 5,534 ops/sec ±1.76% (84 runs sampled)

```