@lencioni 2018-03-03
I think I have a different method using `Object.create(null)` and a reducer that has similar (and potentially better) perf characteristics than `Set`, but without being a breaking change. Here's the results from the css-in-js-perf-tests.

## 1.2.5

```
Running simple test.

  1 test completed.

  aphrodite x 24,003 ops/sec ±1.19% (82 runs sampled)


Running nested test.

  1 test completed.

  aphrodite x 10,371 ops/sec ±2.66% (78 runs sampled)


Running style overload test.

  1 test completed.

  aphrodite x 3,461 ops/sec ±1.31% (83 runs sampled)


Running classes overload test.

  1 test completed.

  aphrodite x 4,069 ops/sec ±1.85% (79 runs sampled)
```

## master

```
Running simple test.

  1 test completed.

  aphrodite x 23,930 ops/sec ±1.94% (84 runs sampled)


Running nested test.

  1 test completed.

  aphrodite x 10,265 ops/sec ±2.25% (81 runs sampled)


Running style overload test.

  1 test completed.

  aphrodite x 3,433 ops/sec ±2.43% (80 runs sampled)


Running classes overload test.

  1 test completed.

  aphrodite x 3,976 ops/sec ±2.68% (81 runs sampled)
```

## Set

```
Running simple test.

  1 test completed.

  aphrodite x 25,067 ops/sec ±1.76% (83 runs sampled)


Running nested test.

  1 test completed.

  aphrodite x 10,425 ops/sec ±2.05% (82 runs sampled)


Running style overload test.

  1 test completed.

  aphrodite x 3,513 ops/sec ±2.00% (79 runs sampled)


Running classes overload test.

  1 test completed.

  aphrodite x 4,311 ops/sec ±2.27% (83 runs sampled)
```

## Object.create

```
Running simple test.

  1 test completed.

  aphrodite x 24,126 ops/sec ±1.88% (83 runs sampled)


Running nested test.

  1 test completed.

  aphrodite x 10,331 ops/sec ±1.73% (83 runs sampled)


Running style overload test.

  1 test completed.

  aphrodite x 3,397 ops/sec ±2.02% (81 runs sampled)


Running classes overload test.

  1 test completed.

  aphrodite x 4,057 ops/sec ±2.27% (82 runs sampled)
```

Profiling in Chrome with a pretty naive benchmark script, it looks like the `Object.create` approach is marginally faster than `Set`. I'll push up a commit soon with this change.